requirements.txt
awsebcli==3.13.0
blessed==1.14.2
boto3==1.7.24
botocore==1.10.24
cement==2.8.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.7
defusedxml==0.5.0
Django==1.11.13
django-admin-honeypot==1.1.0
django-allauth==0.36.0
django-authtools==1.6.0
django-model-utils==3.1.2
django-redis==4.9.0
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
docker==2.6.1
docker-pycreds==0.2.3
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
idna==2.6
jmespath==0.9.3
oauthlib==2.0.7
pathspec==0.5.5
psycopg2==2.7.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.47.0
